# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "description": "API da PoC de crédito interfinanceiro overnight. Com DATABASE_URL, /api/* lê e grava no PostgreSQL (x-data-source: postgres); sem ela, em development/test, usa dados em memória (x-data-source: mock). Os dados do seed e da simulação são sintéticos: hashes e endereços não existem na Sepolia. Erros usam Problem Details. Somente development/test expõem comandos simulados /api/mock e intenções de assinatura, sem autenticação de carteira. Nenhuma rota assina ou envia transações. Guia em docs/api.md.",
    "license": {
      "name": "MIT"
    },
    "title": "API do Backend Banco Inter",
    "version": "0.4.0"
  },
...
  "paths": {
...
    "/openapi.json": {
      "get": {
        "operationId": "getOpenApi",
        "summary": "Contrato OpenAPI",
        "tags": [
          "documentação"
        ],
        "description": "Documento OpenAPI 3.0 desta versão.",
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object"
                }
              }
            }
          },
          "500": {
            "description": "Erro no formato RFC 9457 (Problem Details), servido como application/problem+json.",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "description": "Erro no formato RFC 9457 (Problem Details), servido como application/problem+json.",
                  "properties": {
                    "correlationId": {
                      "description": "Id da requisição, igual ao registrado no log do servidor.",
                      "type": "string"
                    },
                    "detail": {
                      "description": "Explicação legível do erro.",
                      "type": "string"
                    },
                    "status": {
                      "description": "Código HTTP repetido no corpo.",
                      "type": "integer"
                    },
                    "title": {
                      "description": "Resumo curto e estável do erro.",
                      "type": "string"
                    },
                    "type": {
                      "description": "URI estável do tipo de erro. Os tipos possíveis estão em docs/api.md; todos começam com https://api.example.invalid/problems/.",
                      "type": "string"
                    }
                  },
                  "required": [
                    "type",
                    "title",
                    "status",
                    "detail",
                    "correlationId"
                  ],
                  "type": "object"
                }
              }
            }
          }
        }
      }
    }
  },
  "servers": [
    {
      "description": "Local",
      "url": "http://127.0.0.1:3000"
    }
  ],
...
}